finished added graphs to teams page.
